add getTimestampPairs
